Adds context menu in review and comment thread
