+++ src/components/Model/ModelDiscussion/CommentDiscussionItem.tsx
@@ -1,28 +1,23 @@
-import { ActionIcon, Badge, Button, Card, Group, Menu, Text } from '@mantine/core';
-import { closeAllModals, openConfirmModal } from '@mantine/modals';
+import { Badge, Button, Card, Group, Text } from '@mantine/core';
 import { ReviewReactions } from '@prisma/client';
-import { IconDotsVertical, IconTrash, IconEdit, IconFlag, IconMessageCircle2 } from '@tabler/icons';
+import { IconMessageCircle2 } from '@tabler/icons';
 
 import { ContentClamp } from '~/components/ContentClamp/ContentClamp';
-import { LoginRedirect } from '~/components/LoginRedirect/LoginRedirect';
+import { DaysFromNow } from '~/components/Dates/DaysFromNow';
+import { CommentDiscussionMenu } from '~/components/Model/ModelDiscussion/CommentDiscussionMenu';
 import { ReactionPicker } from '~/components/ReactionPicker/ReactionPicker';
 import { RenderHtml } from '~/components/RenderHtml/RenderHtml';
 import { UserAvatar } from '~/components/UserAvatar/UserAvatar';
 import { useCurrentUser } from '~/hooks/useCurrentUser';
 import { useRoutedContext } from '~/routed-context/routed-context.provider';
 import { ReactionDetails } from '~/server/selectors/reaction.selector';
-import { ReportEntity } from '~/server/schema/report.schema';
 import { CommentGetAllItem } from '~/types/router';
-import { showErrorNotification } from '~/utils/notifications';
 import { abbreviateNumber } from '~/utils/number-helpers';
 import { trpc } from '~/utils/trpc';
-import { DaysFromNow } from '~/components/Dates/DaysFromNow';
 
 export function CommentDiscussionItem({ comment }: Props) {
   const { openContext } = useRoutedContext();
   const currentUser = useCurrentUser();
-  const isOwner = currentUser?.id === comment.user.id;
-  const isMod = currentUser?.isModerator ?? false;
 
   const { data: reactions = [] } = trpc.comment.getReactions.useQuery(
     { commentId: comment.id },
@@ -35,36 +30,6 @@ export function CommentDiscussionItem({ comment }: Props) {
   const { data: model } = trpc.model.getById.useQuery({ id: comment.modelId });
 
   const queryUtils = trpc.useContext();
-  const deleteMutation = trpc.comment.delete.useMutation({
-    async onSuccess() {
-      await queryUtils.comment.getAll.invalidate();
-      closeAllModals();
-    },
-    onError(error) {
-      showErrorNotification({
-        error: new Error(error.message),
-        title: 'Could not delete comment',
-      });
-    },
-  });
-  const handleDeleteComment = () => {
-    openConfirmModal({
-      title: 'Delete Comment',
-      children: (
-        <Text size="sm">
-          Are you sure you want to delete this comment? This action is destructive and cannot be
-          reverted.
-        </Text>
-      ),
-      centered: true,
-      labels: { confirm: 'Delete Comment', cancel: "No, don't delete it" },
-      confirmProps: { color: 'red', loading: deleteMutation.isLoading },
-      closeOnConfirm: false,
-      onConfirm: () => {
-        deleteMutation.mutate({ id: comment.id });
-      },
-    });
-  };
 
   const toggleReactionMutation = trpc.comment.toggleReaction.useMutation({
     async onMutate({ id, reaction }) {
@@ -127,44 +92,7 @@ export function CommentDiscussionItem({ comment }: Props) {
             linkToProfile
           />
         </Group>
-        <Menu position="bottom-end" withinPortal>
-          <Menu.Target>
-            <ActionIcon size="xs" variant="subtle">
-              <IconDotsVertical size={14} />
-            </ActionIcon>
-          </Menu.Target>
-          <Menu.Dropdown>
-            {(isOwner || isMod) && (
-              <>
-                <Menu.Item
-                  icon={<IconTrash size={14} stroke={1.5} />}
-                  color="red"
-                  onClick={handleDeleteComment}
-                >
-                  Delete comment
-                </Menu.Item>
-                <Menu.Item
-                  icon={<IconEdit size={14} stroke={1.5} />}
-                  onClick={() => openContext('commentEdit', { commentId: comment.id })}
-                >
-                  Edit comment
-                </Menu.Item>
-              </>
-            )}
-            {(!currentUser || !isOwner) && (
-              <LoginRedirect reason="report-model">
-                <Menu.Item
-                  icon={<IconFlag size={14} stroke={1.5} />}
-                  onClick={() =>
-                    openContext('report', { type: ReportEntity.Comment, entityId: comment.id })
-                  }
-                >
-                  Report
-                </Menu.Item>
-              </LoginRedirect>
-            )}
-          </Menu.Dropdown>
-        </Menu>
+        <CommentDiscussionMenu comment={comment} user={currentUser} />
       </Group>
 
       <ContentClamp maxHeight={100}>
